:lipstick: Style: 기본 설명 페이지와 코스 상세 페이지 분리
- placeInfo related_course에 따라 구분
- placeInfo용 type 추가
