+++ src/components/ExplainPage/CourseExplain.tsx
@@ -1,0 +1,5 @@
+const CourseExplain = () => {
+  return <section>코스 관련 상세 페이지 컴포넌트</section>;
+};
+
+export default CourseExplain;
